massa_grpc

Module private

business code for node management methods

Functions

add_staking_secret_keys 🔒

Add staking secret keys to wallet

add_to_bootstrap_blacklist 🔒

Add IP addresses to node bootstrap blacklist

add_to_bootstrap_whitelist 🔒

Add IP addresses to node bootstrap whitelist

add_to_peers_whitelist 🔒

Add IP addresses to node peers whitelist. No confirmation to expect. Note: If the ip was unknown it adds it to the known peers, otherwise it updates the peer type

allow_everyone_to_bootstrap 🔒

Allow everyone to bootstrap from the node by removing bootstrap whitelist configuration file

ban_nodes_by_ids 🔒

Ban multiple nodes by their individual ids

ban_nodes_by_ips 🔒

Ban multiple nodes by their individual IP addresses

get_bootstrap_blacklist 🔒

Get node bootstrap blacklist IP addresses

get_bootstrap_whitelist 🔒

Get node bootstrap whitelist IP addresses

get_mip_status 🔒

get_node_status 🔒

Get node status

get_peers_whitelist 🔒

Get node peers whitelist IP addresses

remove_from_bootstrap_blacklist 🔒

Remove from bootstrap blacklist given IP addresses

remove_from_bootstrap_whitelist 🔒

Remove from bootstrap whitelist given IP addresses

remove_from_peers_whitelist 🔒

Remove from peers whitelist given IP addresses

remove_staking_addresses 🔒

Remove addresses from staking

shutdown_gracefully 🔒

Shutdown the node gracefully

sign_messages 🔒

Sign messages with node’s key

unban_nodes_by_ids 🔒

Unban multiple nodes by their individual ids

unban_nodes_by_ips 🔒

Unban multiple nodes by their individual IP addresses